requirements.txt
altgraph==0.17.4
macholib==1.16.3
packaging==24.2
pillow==11.0.0
qrcode==8.0
